🚐 move hooks directory to top level server folder
